feat: Adding Nav Menu
So we add the ability to return home by popping a nav item in the header in the layout.
Then we need to completely refactor the project as that layout currently also adds the same header to Sanity Studio! It's interesting how we can structure the project to manage this.